fix(app): register the cobbler-logo icon before it's first used
MatIconRegistry.addSvgIcon('cobbler-logo', ...) lived in
NavbarComponent's constructor, but ManageMenuComponent's own template
uses the same icon earlier in the render tree than the embedded
<cobbler-navbar> (Angular initializes components in template order).
That earlier usage failed with "Unable to find icon with the name
':cobbler-logo'" since the registration hadn't happened yet, even
though NavbarComponent's own icon usage (registered by the time it
renders) worked fine and could mislead you into thinking it was a
fetch/network problem.

Moved the registration to AppComponent's constructor, the actual root
component that always initializes before any of its children. Also
bundles the icon's SVG locally (assets/images/cobbler-logo.svg)
instead of fetching it from cobbler.github.io at runtime, removing an
unrelated but real hard dependency on an external site being
reachable.
