+++ projects/cobbler-frontend/src/app/navbar/navbar.component.spec.ts
@@ -1,6 +1,7 @@
 import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { MatIconModule } from '@angular/material/icon';
+import { MatIconTestingModule } from '@angular/material/icon/testing';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { COBBLER_URL } from 'cobbler-api';
@@ -22,6 +23,7 @@ describe('NavbarComponent', () => {
     await TestBed.configureTestingModule({
       imports: [
         MatIconModule,
+        MatIconTestingModule,
         MatToolbarModule,
         MatSnackBarModule,
         MatButtonModule,
